Improve benchmark tooling
